// generated/src/app/api-models/pcpaa-rules-dtl.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PcpaaRulesDtl {
  ruleId: string;
  seqAttrbId: number;
  urbanToMiles: number;
  ruralToMiles: number;
  operator: string;
  userdefValue: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}